requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=1.55.0
faiss-cpu>=1.9.0
python-dotenv>=1.0.1
pydantic>=2.9.0
pydantic-settings>=2.6.0
numpy>=1.26.0
httpx>=0.27.0
requests>=2.32.0
pytest>=8.3.0
